requirements.txt
# Core
pytz==2018.7
Django==2.1.7
django-configurations==2.1
gunicorn==19.9.0
newrelic==4.10.0.112
coreapi==2.3.3
requests==2.21.0

# For the persistence stores
psycopg2-binary==2.7.6.1
#psycopg2==2.7.6
dj-database-url==0.5.0

# Model Tools
django-model-utils==3.1.2
django_unique_upload==0.2.1

# Rest apis
djangorestframework==3.9.2
Markdown==3.0.1
django-filter==2.0.0
django_extensions==2.1.2

# Developer Tools
ipdb==0.11
ipython==7.2.0
mkdocs==1.0.4
flake8==3.6.0

# Testing
mock==2.0.0
factory-boy==2.11.1
django-nose==1.4.6
nose-progressive==1.5.2
coverage==4.5.2

# Static and Media Storage
django-storages==1.7.1
boto==2.49.0
boto3==1.9.70

# Stripe
stripe==2.17.0

# ImageField
Pillow==5.4.1

# Email
django-anymail[amazon_ses]==5.0

# Debug Toolbar
django-debug-toolbar==1.11

# Admin QoL
# django-nested-admin==3.0.21

# Django custom admin
# django-jet
# django-suit==0.2.25
# https://github.com/darklow/django-suit/tarball/v2
# django-grappelli

# Django coupons
# django-coupons
# test
# -e git+https://github.com/Amurmurmur/django-coupons#egg=django-coupons

#Sentry
sentry-sdk

celery==4.2.1
django-celery-results==1.0.4
django-celery-beat==1.4.0
redis==3.2.1
googlemaps==3.0.2
pycurl==7.43.0.2

prices==1.0.0

# Calculate distance
# mpu==0.17.0
geopy==1.18.1

# Export as CSV
django-import-export==1.1.0

# Geo fields maps
django-map-widgets

# Geo forms for PylogonFields
# django-floppyforms


#Django-cron
#django_cron
#django-crontab
google-measurement-protocol==1.0.0
maxminddb-geolite2==2018.703
django-versatileimagefield==1.10
# Django countries
django-countries==5.3.3
django-prices-openexchangerates==1.0.1
django-prices-vatlayer==1.0.2
django-prices==1.0.4
django-measurement==3.1.1
measurement==2.0.1
#-e git@github.com:celery/django-celery-beat.git#egg=django-celery-beat
# git+git://github.com/mozilla/elasticutils.git#egg=elasticutils
#-e git+https://github.com/celery/django-celery-beat#egg=django-celery-beat